+++ gary/llm/randy.py
